requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-jsonfield==1.4.1
djangorestframework==3.13.1
idna==3.3
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
